+++ api/_routers/97-optional-access-token.go
@@ -7,7 +7,6 @@ import (
 	"github.com/getsentry/sentry-go"
 	"github.com/sirupsen/logrus"
 	"github.com/t2bot/matrix-media-repo/api/_apimeta"
-	"github.com/t2bot/matrix-media-repo/api/_auth_cache"
 	"github.com/t2bot/matrix-media-repo/api/_responses"
 	"github.com/t2bot/matrix-media-repo/common/config"
 	"github.com/t2bot/matrix-media-repo/common/rcontext"
@@ -27,14 +26,15 @@ func OptionalAccessToken(generator GeneratorWithUserFn) GeneratorFn {
 		}
 		if config.Get().SharedSecret.Enabled && accessToken == config.Get().SharedSecret.Token {
 			ctx = ctx.LogWithFields(logrus.Fields{"sharedSecretAuth": true})
-			return generator(r, ctx, _apimeta.UserInfo{
+			ctx, user := applyTrustedRepoAdminHeader(r, ctx, _apimeta.UserInfo{
 				UserId:      "@sharedsecret",
 				AccessToken: accessToken,
 				IsShared:    true,
 			})
+			return generator(r, ctx, user)
 		}
 		appserviceUserId := util.GetAppserviceUserIdFromRequest(r)
-		userId, isGuest, err := _auth_cache.GetUserId(ctx, accessToken, appserviceUserId)
+		userId, isGuest, err := getUserIdForToken(ctx, accessToken, appserviceUserId)
 		if isGuest {
 			return _responses.GuestAuthFailed()
 		}
@@ -50,10 +50,14 @@ func OptionalAccessToken(generator GeneratorWithUserFn) GeneratorFn {
 		}
 
 		ctx = ctx.LogWithFields(logrus.Fields{"authUserId": userId})
-		return generator(r, ctx, _apimeta.UserInfo{
+		user := _apimeta.UserInfo{
 			UserId:      userId,
 			AccessToken: accessToken,
 			IsShared:    false,
-		})
+		}
+		if userId != "" {
+			ctx, user = applyTrustedRepoAdminHeader(r, ctx, user)
+		}
+		return generator(r, ctx, user)
 	}
 }